Search & Replace dialog: when you hit "ESC" the window is hidden. Corrected button sensitiveness problem.
'SeachComponent' class was created and now aggregates all code related
to the Search dialog.

The Search dialog is now persistent and has context sensitivity. When 
you click the 'Escape' key the dialog is hidden. When you click the 
'Find' button the window is no longer hidden.

Added one more event 'search-event' to the 'EditWindow' class.

Added tooltips to the toolbar buttons.

Renamed the method 'get_current_buffer' to 'get_buffer' and added a 
property named 'property' that returns the current buffer.