The method used to target new buffers wasn't working properly. Now CulebraBuffer has a new field 'is_new' and the EditWindow has a 'count_new' to better track the new buffers.
Added commands sensitivity to 'FileSave', 'FileSaveAs', 'EditUndo' and 
'EditRedo' actions.

On '_new_tab_method' changed the algorith to calculate if there is a 
filename to a more clear one.

Now only replaces a new buffer for a opened one if it's the only one and 
if it's unmodified.

Changed the 'CulebraView' constructure to not generate a 'CulebraBuffer'
by instance and not only once.

The 'EditWindow' now contains a 'buffer-changed' and is tracked by
'culebra.plugin.Plugin'.