Made the auto-complete feature a culebra option and disabled it by default (IMO it should be hidden in the next release, since it's broken).
Added syntatic sugar 'PathComponent' and 'PropertyWrapper', the first is 
based on PEAK the second is a simplification of a design pattern. These 
functions should be moved to the 'components.py' module.

Made the Find and Replace dialogs non-modal and corrected how Escape 
hides them.

Connected the dialog button on the Find dialog as a proxy to the 
"EditFindNext" action.

Corrected a bug introduced while refactoring in the Find dialog.